requirements.txt
streamlit>=1.0.0
google-generativeai>=0.3.0
SpeechRecognition>=3.8.1
pyttsx3>=2.90